refactor(bin): split memory dump into separate plugin
This means we shouldn't need one giant executable with tons of QML subdirs.
