+++ ui/CMakeLists.txt
@@ -1,1 +1,2 @@
 add_subdirectory(highlight)
+add_subdirectory(memory)
