+++ ui/memory/hexdump/MemoryDumpBorder.qml
