+++ ui/memory/hexdump/MemoryDumpEdit.qml
