+++ ui/memory/plugin.cpp
@@ -1,0 +1,3 @@
+#include "plugin.hpp"
+
+void Memory::registerTypes(const char *uri) {}
